feat: implement datum either adt with eitherT
Trying out an unsafe either monad transformer with the $
substitution type. It passes the monad laws tests and the
surface area of the implementation only relies on getting
the composition of chain and of correct, since the rest of
the monad functions are derived.

I also made the getSemigroup for either more robust. It
takes a semigroup for both left and right. It will always
favor right, but will concat left values if possible.

Added another monad transformer for option, but don't have
a use for it yet.

Removed unused imports where I saw them.

Changed deno test name for adt modules to {MODULE_NAME} Modules