feat: inncrease type accuracy of record traverse
The traverse type for record was not useful as a pipeable
operator as it was losing both the key structure and the
output type of the record.

To fix this I altered the structure of pipeable traversable
to by fully curried across three arguments. I also had to
update the derivation function for pipeable traversable. This
made traverse properly infer the inputs and outputs during
inline piping.

The second issue was keeping the key structure for traverse
and map. In this case, since the pipeables are meant to be
used as standalone functions, I reexported them with altered
signatures that properly keep the key structure when being
piped. Rather than wrap the derived functions in a new
function definition, I reexported with a coerced type. I
might consider changing this in the future as I aim to
avoid type coercion where possible.