feat: much work towards io-ts
It was becoming difficult to follow the logic of the Kliesli
decoder implementation in the new io-ts. I found it easier,
if a little uglier, to implement Decoder with brute force in
isolation. I found that my implementation of getRightBifunctor
in either.ts was flawed so it was removed.

Following are new features that were ported from fp-ts or io-ts:
* initial decoder
* array: getMonoid and getShow
* bifunctor: added mapLeft
* derivations: added createBifunctor and createPipeableBifunctor
* either: use createBifunctor and createPipeableBifunctor
* schemable: simplify types for sum and intersect, add sections
* guard: add TypeOf, InputOf, and new sum/intersect
* hkts: internal only testing for cleaner type messages
* record: export optimization fns and more indexed types for traverse
* tree: minimal port of tree from fp-ts
* type_classes: make all modules readonly, minor formatting