feat: more port of monocle-ts
Cleaned up some types in the fns.ts file. Also added the flow
operator as it's used in the monocle-ts port.

Implemented the following basic monocle-ts files:
* iso
* lens
* optional
* prism

I haven't implemented traversal yet, but will soon.

I also updated getOrElse in option.ts to be pipeable and
renamed getMonad to getRightMonad in these.ts so it is
more accurate. This necessitated updating a few references.