feat: abstract assertMonad and refactor io
Previously, the monad, applicative, functor, and apply
laws were mostly tested by an assertMonad function. Unfortunately,
this function could not automate the testing of adts like
reader, io, task, and their composed forms.

assertMonad and its dependencies were extended to take a run
parameter that "evaluates" the resultant adt into a plain object
wrapped in a promise that can be compared using assertEquals.
This allows the use of the assertMonad test for the asynchronous
adts.

Also, some minor "optimizations" were implemented for io and
ioEither