feat: test refactor and sequence fix
The initial typeclass tests were difficult to use and limited
in their effectiveness. I added tests for all static-land type
classes except chainrec and traversable such that they have
similar signatures and require passing instance variables so
multiiple law paths can be tested for a single type class
implementation.

I migrated the existing adts to use the new tests. Moving
forward I need to implement all modules that are possible and
start testing the module getters as well.

Additionally, there was a serious bug in the implementation
of sequence that showed up when I was playing with sequenceTuple
for the Array Apply. I refactored the tuple and record constructors
in sequence.ts to fix the bug and implemented sequenceTuple and
sequenceStruct for array.ts and tree.ts.

While building out the new test asserts I found that semigroupoid,
category, and profuctor were not parameterized over type
constructor length, so I implemented that as well. This allowed
me to remove some type casting from the optics files where id
was implemented for category.