feat: widen pipeable chain
While I'm ver averse to widening the type of the chain defined
in the Chain static-land type class, I have no problem making
the pipeable type class functions more useful. So for chain I
widened the type of the 2nd, 3rd, and 4th type parameters.

In the ongoing effort to both optimize and dogfood I started
using array IndexedTraversal from array.ts and record.ts in
decoder.ts. There are about the same number of any's as before
and this work will need to be revisited to tighten up the
types.

In general I'm not a fan of the M, T, C, and W suffixes that
haskell uses for go knows why. Often I see these being
replicated in fp-ts and I fell prey to using them for
getEitherM and getOptionM here. To rectify this I switched
both usages to the more illustrative composeMonad name. This
required updating their references in io_either.ts and
task_either.ts.

I also cleaned up some formatting in record.ts and have taken
to using deno-fmt-ignore to make the indexed types a bit
more readable.