fix: initialize Applicative of types with never
When second through fourth type parameters in the of
function are not given an initial type they revert to
unknown, which will consume all other types in a union.
This causes some issues for the pipeable operators with
type widening. Like the type widening, this was not
applied to the static-land ttype class types, only the
ergonomic P types.