feat: implement encoder and codec
Revisted the io-ts port again (likely not for the last time).
I implemented encoder.ts and codec.ts. I also found that there
were a few typing issues with Schema and the make utility so
I resolved those for now.

Cleaned up some of the organization. This included moving
memoize, typeOf, and intersect to the fns.ts file.

Since I'm not implementing Kliesli and am deferring to the
simpler base implementations for Decoder, Encoder, and Codec
I opted to remove the builder combinators fromType, fromPartial
etc, instead preferring the Schemable functions. This cuts
down on the complexity of the implementations substantially.

Also, rather than compacting a record for the partial
Schemable I created an undefinable combinator that mirrors
the nullable one. This combinator seemed more appropriate
given the intention of partial.

Lastly, I cleaned up some of the TODO items and expanded
on their tasks.