feat: finish Traversal implementation
Needed to add some pipeable functions to array.ts and record.ts
for easy lookup, insert and delete on those objects. Also, since
the Index and At types are related to Optional and Lens
respectively (they return those instances) I added those types
and a couple of constructors to the related files. This allowed
my to implement index, key, and atKey combinators for Traversal.

I should also go back and implement the index, key, and atKey
combinators for the other optics.