fix: get deno lint to pass
Although there are some cases where any is an acceptable type,
as in the extends clause of type narrowing, this commit
endeavors to remove as much as possible. Occasionally I have
had to move to a cast to unknown and then the associated
type.

This ended up exposing a bug in composeMonad for Option,
which was fixed. I'll need to come back to these lint rules
at some point in the future.

As a note, I am still dissatisfied with the way composeMonad
is implemented typewise. I think it will be necessary to
rewrite those by hand at some point instead of relying on
the createMonad derivative function.

I've updated the github action to run deno lint in addition
too the tests.