feat: implement experimental nilable.ts
Option is a fantastic algebraic data type. Unfortunately, it
comes with the price of wrapping all values in a new object.
I've implemented an idea similar to Option that creates
static land modules for the type Nil<T> = T | undefined | null.

Maybe this is folly, but I'll try it out and see.