feat: implement traverse for all optics
In order to implement traverse without merging all work into
an "internal" file or cause cyclic import problems I needed to
hoist fromTraverable to the derivations.ts file. It's starting
to make sense why gcanti did most work in the internal file, but
I'm confident that implementing all of the composition functions
as well as the converters and pipeable was worth it for learning
purposes.