feat: revisit schemable and decoder
The interfaces for decoder, encoder, codec, guard, and schemable
came from the experimental io-ts work done by gcanti.

I believe originally he broke out the decode, encode, and guard
functions so the could be combined like other type class modules
into the same object. After much thought I decided that due to the
new Schemaable and Schema types it made more sense to get rid
of the extra objects and to represent decoder and guard more
directly. This makes the "blessed" interface for these the one
that uses Schemable.make and then uses the schemabele consts from
those modules to build them. ie. write one type and use it to
generate the others.. Or use the others directly. The Codec
abstraction seemed pointless to me, as did encode, since they
cannot represent external types anyway.

That said, I also implemented a Schemable for JsonSchema. This
means that a json schema can be generated from the same Schema
that generates a decoder. I believe that the Schemable type
should also be extennded to include a few more JsonSchema
features that we can implement in decoder and guard as well.

All in all, here are the following changes:
1. Implement json_schema.ts
2. Implement reader_either.ts
3. Remove encoder.ts
4. Remove codec.ts
5. Alter Guard to be an alias of Refinement
6. Alter Decoder to be an alias of (i: I) => Either<E, A>