feat: major structural changes
The async and "magic" module testing for state, reader, etc
from the assert.ts lib was not really useful in the long run.

I removed all of the async and await from assert.ts and updated
the tests where applicable. In cases like task, task_either,
state, and reader, I removed the law testing altogether and
will need to write manual tests for those adts.

This commit implements a break with the static-land spec. After
noticing that my speed of implementing new features was starting
to drag I started tracking where I was spending my time. The
primary culprit was implementing static land specs and immediately
setting them up to be curried.

Thus, I've removed all of the curried static-land types and moved
everything to use curried or partially curried types. This turned
out to simplify tons of stuff, from optics implementations to
testing, there is just less to worry about. Since one of my minor
goals with hkts is to reduce the amount of complelxity in the
type system this seems like its a good move in that direction.

In addition to this break I've removed the composeMonad functions
in a few places, since implementing the monad interface is
often easier and more optimized without the compositions.

* array.ts tests and getSetoid for graph.ts
* finished const tests
* added tests for datum.ts
* createMonad takes an optional map for optimization
* add swap fn to fns.ts
* implemented initial graph.ts like algebraic graphs
* implemented initial map.ts like fp-ts
* implemented initial monoid.ts like fp-ts
* added exists combinator to option.ts
* moved all optics into optics folder
* added compare with ordering to ord.ts
* updated types.ts substitution to handle Map and Set
* implemented initial set.ts like fp-ts
* moved decoder, guard, and schemable to schemable folder
* started filling out array.test.ts tests